+++ elastic-job-common/elastic-job-common-core/src/test/java/com/dangdang/ddframe/job/util/AllUtilTests.java
@@ -26,8 +26,6 @@ import com.dangdang.ddframe.job.util.env.LocalHostServiceTest;
 import com.dangdang.ddframe.job.util.env.TimeServiceTest;
 import com.dangdang.ddframe.job.util.json.GsonFactoryTest;
 import com.dangdang.ddframe.job.util.json.JobConfigurationGsonTypeAdapterTest;
-import com.dangdang.ddframe.job.util.restful.RestfulExceptionTest;
-import com.dangdang.ddframe.job.util.restful.RestfulServerTest;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.junit.runner.RunWith;
@@ -43,8 +41,6 @@ import org.junit.runners.Suite.SuiteClasses;
         HostExceptionTest.class, 
         GsonFactoryTest.class, 
         JobConfigurationGsonTypeAdapterTest.class, 
-        RestfulServerTest.class, 
-        RestfulExceptionTest.class, 
         ShardingItemsTest.class, 
         ShardingItemParametersTest.class
     })